+++ examples/fishjam-chat/components/VideosGrid/parsePeersToTracks.ts
@@ -1,0 +1,41 @@
+import { PeerWithTracks } from '@fishjam-cloud/react-native-client/build/hooks/usePeers';
+import { GridTrack } from './GridTrackItem';
+import { PeerMetadata } from '../../types/metadata';
+
+const createGridTrackFromPeer = (
+  peer: PeerWithTracks<PeerMetadata>,
+): GridTrack[] => {
+  const tracks: GridTrack[] = [];
+
+  if (peer.cameraTrack && peer.cameraTrack.isActive) {
+    tracks.push({
+      ...peer.cameraTrack,
+      peerId: peer.id,
+      isLocal: peer.isLocal,
+      userName: peer.metadata.peer.displayName,
+      isVadActive: peer.microphoneTrack?.vadStatus === 'speech',
+    });
+  }
+
+  if (peer.screenShareVideoTrack && peer.screenShareVideoTrack.isActive) {
+    tracks.push({
+      ...peer.screenShareVideoTrack,
+      peerId: peer.id,
+      isLocal: peer.isLocal,
+      userName: peer.metadata.peer.displayName,
+      isVadActive: peer.screenShareAudioTrack?.vadStatus === 'speech',
+    });
+  }
+
+  return tracks;
+};
+
+export const parsePeersToTracks = (
+  localPeer: PeerWithTracks<PeerMetadata> | null,
+  remotePeers: PeerWithTracks<PeerMetadata>[],
+): GridTrack[] => {
+  return [
+    ...(localPeer ? createGridTrackFromPeer(localPeer) : []),
+    ...remotePeers.flatMap(createGridTrackFromPeer),
+  ];
+};
